# evm: non-interface IntraBlockState