--- a/core/vm/evm_test.go
+++ b/core/vm/evm_test.go
@@ -28,10 +28,16 @@ import (
 )
 
 func TestInterpreterReadonly(t *testing.T) {
+	//tx, sd := testTemporalTxSD(t, testTemporalDB(t))
+	//defer tx.Rollback()
+	//
+	//r, w := state.NewReaderV3(sd), state.NewWriter(sd, nil)
+	//s := state.New(r)
+
 	t.Parallel()
 	c := NewJumpDestCache(128)
 	rapid.Check(t, func(t *rapid.T) {
-		env := NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, &dummyStatedb{}, chain.TestChainConfig, Config{})
+		env := NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, nil, chain.TestChainConfig, Config{})
 
 		isEVMSliceTest := rapid.SliceOfN(rapid.Bool(), 1, -1).Draw(t, "tevm")
 		readOnlySliceTest := rapid.SliceOfN(rapid.Bool(), len(isEVMSliceTest), len(isEVMSliceTest)).Draw(t, "readonly")
@@ -291,7 +297,7 @@ func TestReadonlyBasicCases(t *testing.T) {
 				t.Parallel()
 				readonlySliceTest := testcase.readonlySliceTest
 
-				env := NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, &dummyStatedb{}, chain.TestChainConfig, Config{})
+				env := NewEVM(evmtypes.BlockContext{}, evmtypes.TxContext{}, nil, chain.TestChainConfig, Config{})
 
 				readonliesGot := make([]*readOnlyState, len(testcase.readonlySliceTest))
 				isEVMGot := make([]bool, len(evmsTestcase.emvs))