--- a/core/vm/evm_test.go
+++ b/core/vm/evm_test.go
@@ -20,11 +20,12 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/holiman/uint256"
+	"pgregory.net/rapid"
+
 	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon/core/vm/evmtypes"
-	"github.com/holiman/uint256"
-	"pgregory.net/rapid"
 )
 
 func TestInterpreterReadonly(t *testing.T) {