--- a/core/vm/evmtypes/evmtypes.go
+++ b/core/vm/evmtypes/evmtypes.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/holiman/uint256"
 
-	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/types"
 	"github.com/erigontech/erigon/core/tracing"
@@ -120,60 +119,9 @@ type (
 
 // IntraBlockState is an EVM database for full state querying.
 type IntraBlockState interface {
-	CreateAccount(common.Address, bool) error
-
 	SubBalance(common.Address, *uint256.Int, tracing.BalanceChangeReason) error
 	AddBalance(common.Address, *uint256.Int, tracing.BalanceChangeReason) error
 	GetBalance(common.Address) (*uint256.Int, error)
-
-	GetNonce(common.Address) (uint64, error)
-	SetNonce(common.Address, uint64) error
-
-	GetCodeHash(common.Address) (common.Hash, error)
-	GetCode(common.Address) ([]byte, error)
-	SetCode(common.Address, []byte) error
-	GetCodeSize(common.Address) (int, error)
-
-	// eip-7702; delegated designations
-	ResolveCodeHash(common.Address) (common.Hash, error)
-	ResolveCode(common.Address) ([]byte, error)
-	GetDelegatedDesignation(common.Address) (common.Address, bool, error)
-
-	AddRefund(uint64)
-	SubRefund(uint64)
-	GetRefund() uint64
-
-	GetCommittedState(common.Address, common.Hash, *uint256.Int) error
-	GetState(address common.Address, slot common.Hash, outValue *uint256.Int) error
-	SetState(common.Address, common.Hash, uint256.Int) error
-
-	GetTransientState(addr common.Address, key common.Hash) uint256.Int
-	SetTransientState(addr common.Address, key common.Hash, value uint256.Int)
-
-	Selfdestruct(common.Address) (bool, error)
-	HasSelfdestructed(common.Address) (bool, error)
-	Selfdestruct6780(common.Address) error
-
-	// Exist reports whether the given account exists in state.
-	// Notably this should also return true for suicided accounts.
-	Exist(common.Address) (bool, error)
-	// Empty returns whether the given account is empty. Empty
-	// is defined according to EIP161 (balance = nonce = code = 0).
-	Empty(common.Address) (bool, error)
-
-	Prepare(rules *chain.Rules, sender, coinbase common.Address, dest *common.Address,
-		precompiles []common.Address, txAccesses types.AccessList, authorities []common.Address) error
-
-	AddressInAccessList(addr common.Address) bool
-	// AddAddressToAccessList adds the given address to the access list. This operation is safe to perform
-	// even if the feature/fork is not active yet
-	AddAddressToAccessList(addr common.Address) (addrMod bool)
-	// AddSlotToAccessList adds the given (address,slot) to the access list. This operation is safe to perform
-	// even if the feature/fork is not active yet
-	AddSlotToAccessList(addr common.Address, slot common.Hash) (addrMod, slotMod bool)
-
-	RevertToSnapshot(int)
-	Snapshot() int
 
 	AddLog(*types.Log)