--- a/core/vm/interface.go
+++ b/core/vm/interface.go
@@ -22,12 +22,11 @@ package vm
 import (
 	"math/big"
 
-	"github.com/erigontech/erigon/core/state"
 	"github.com/holiman/uint256"
 
 	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/common"
-
+	"github.com/erigontech/erigon/core/state"
 	"github.com/erigontech/erigon/core/vm/evmtypes"
 )