--- a/eth/tracers/logger/logger_test.go
+++ b/eth/tracers/logger/logger_test.go
@@ -49,18 +49,6 @@ func (d *dummyContractRef) SetBalance(*big.Int)        {}
 func (d *dummyContractRef) SetNonce(uint64)            {}
 func (d *dummyContractRef) Balance() *big.Int          { return new(big.Int) }
 
-type dummyStatedb struct {
-	state.IntraBlockState
-}
-
-func (*dummyStatedb) GetRefund() uint64                                              { return 1337 }
-func (*dummyStatedb) GetState(_ common.Address, _ common.Hash, _ *uint256.Int) error { return nil }
-func (*dummyStatedb) SetState(_ common.Address, _ common.Hash, _ uint256.Int) error  { return nil }
-func (*dummyStatedb) AddSlotToAccessList(_ common.Address, _ common.Hash) (addrMod, slotMod bool) {
-	return false, false
-}
-func (*dummyStatedb) GetCommittedState(common.Address, common.Hash, *uint256.Int) error { return nil }
-
 func TestStoreCapture(t *testing.T) {
 	c := vm.NewJumpDestCache(128)
 	ibs := state.New(state.NewNoopReader())