--- a/rpc/jsonrpc/eth_callMany.go
+++ b/rpc/jsonrpc/eth_callMany.go
@@ -232,7 +232,7 @@ func (api *APIImpl) CallMany(ctx context.Context, bundles []Bundle, simulateCont
 	// after replaying the txns, we want to overload the state
 	// overload state
 	if stateOverride != nil {
-		err = stateOverride.Override((evm.IntraBlockState()).(*state.IntraBlockState))
+		err = stateOverride.Override(evm.IntraBlockState())
 		if err != nil {
 			return nil, err
 		}