Mostly was motivated by seen `AccessList` methods in profiler
```
go test -bench=BenchmarkCall -run=Benchmark  ./core/vm/runtime -count=10 > new2.txt
go test -bench=BenchmarkCall -run=Benchmark  ./core/vm/runtime -count=10 > old2.txt

/home/erigon/go/bin/benchstat old2.txt new2.txt
goos: linux
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: AMD EPYC 4344P 8-Core Processor
        │  old2.txt   │              new2.txt              │
        │   sec/op    │   sec/op     vs base               │
Call-16   3.165m ± 0%   3.097m ± 1%  -2.13% (p=0.000 n=10)

        │   old2.txt   │              new2.txt               │
        │     B/op     │     B/op      vs base               │
Call-16   1.800Mi ± 0%   1.758Mi ± 0%  -2.34% (p=0.000 n=10)

        │  old2.txt   │            new2.txt             │
        │  allocs/op  │  allocs/op   vs base            │
Call-16   19.20k ± 0%   19.20k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

```
go test -bench=BenchmarkSimple -run=Benchmark  ./core/vm/runtime -count=10 > new2.txt
go test -bench=BenchmarkSimple -run=Benchmark  ./core/vm/runtime -count=10 > old2.txt


/home/erigon/go/bin/benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/erigontech/erigon/core/vm/runtime
cpu: AMD EPYC 4344P 8-Core Processor
                                       │   old.txt    │               new.txt               │
                                       │    sec/op    │    sec/op     vs base               │
SimpleLoop/staticcall-identity-100M-16   233.9m ± 12%   231.3m ±  2%       ~ (p=0.190 n=10)
SimpleLoop/call-identity-100M-16         338.1m ±  1%   327.9m ±  2%  -3.03% (p=0.000 n=10)
SimpleLoop/loop-100M-16                  264.7m ±  3%   261.7m ±  3%       ~ (p=0.218 n=10)
SimpleLoop/loop2-100M-16                 305.1m ±  3%   343.2m ± 19%       ~ (p=0.143 n=10)
SimpleLoop/call-nonexist-100M-16         309.5m ±  3%   315.1m ±  5%       ~ (p=0.105 n=10)
SimpleLoop/call-EOA-100M-16              348.1m ±  1%   328.5m ±  2%  -5.63% (p=0.001 n=10)
SimpleLoop/call-reverting-100M-16        503.7m ±  2%   472.6m ±  1%  -6.17% (p=0.000 n=10)
geomean                                  320.4m         318.6m        -0.56%

                                       │   old.txt    │               new.txt               │
                                       │     B/op     │     B/op      vs base               │
SimpleLoop/staticcall-identity-100M-16   67.94Mi ± 0%   67.94Mi ± 0%       ~ (p=0.075 n=10)
SimpleLoop/call-identity-100M-16         66.57Mi ± 0%   66.57Mi ± 0%  -0.00% (p=0.041 n=10)
SimpleLoop/loop-100M-16                  2.105Ki ± 7%   1.967Ki ± 7%  -6.59% (p=0.046 n=10)
SimpleLoop/loop2-100M-16                 2.105Ki ± 0%   2.105Ki ± 7%       ~ (p=0.582 n=10)
SimpleLoop/call-nonexist-100M-16         22.78Mi ± 0%   22.78Mi ± 0%       ~ (p=0.582 n=10)
SimpleLoop/call-EOA-100M-16              22.78Mi ± 0%   22.78Mi ± 0%  -0.00% (p=0.000 n=10)
SimpleLoop/call-reverting-100M-16        201.7Mi ± 0%   201.7Mi ± 0%  -0.00% (p=0.005 n=10)
geomean                                  2.962Mi        2.933Mi       -0.97%

                                       │   old.txt   │               new.txt                │
                                       │  allocs/op  │  allocs/op   vs base                 │
SimpleLoop/staticcall-identity-100M-16   2.740M ± 0%   2.740M ± 0%       ~ (p=0.582 n=10)
SimpleLoop/call-identity-100M-16         2.685M ± 0%   2.685M ± 0%       ~ (p=0.470 n=10)
SimpleLoop/loop-100M-16                   13.00 ± 0%    13.00 ± 0%       ~ (p=1.000 n=10) ¹
SimpleLoop/loop2-100M-16                  13.00 ± 0%    13.00 ± 0%       ~ (p=1.000 n=10) ¹
SimpleLoop/call-nonexist-100M-16         746.3k ± 0%   746.3k ± 0%       ~ (p=0.582 n=10)
SimpleLoop/call-EOA-100M-16              746.3k ± 0%   746.3k ± 0%  -0.00% (p=0.001 n=10)
SimpleLoop/call-reverting-100M-16        3.571M ± 0%   3.571M ± 0%  -0.00% (p=0.048 n=10)
geomean                                  58.94k        58.94k       -0.00%
¹ all samples are equal
```